pulling to my own fork

+++ code/datums/components/rot.dm
@@ -37,6 +37,9 @@
 
 /datum/component/rot/corpse/process()
 	var/mob/living/carbon/C = parent
+	if(C == null) //can't delete what doesnt exist
+		return
+
 	if(C.stat != DEAD)
 		qdel(src)
 		return
